% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_kitty.R
\name{is_kitty}
\alias{is_kitty}
\title{Determine if we are running in a kitty terminal}
\usage{
is_kitty(throw = FALSE)
}
\arguments{
\item{throw}{Throw an error when we are running in a kitty terminal.}
}
\value{
If \code{throw = FALSE}, returns `TRUE` if R is running in a kitty terminal
and `FALSE` otherwise. If, \code{throw = FALSE} will throw an error when not
running in a kitty terminal; returns `TRUE` otherwise.
}
\description{
Determine if we are running in a kitty terminal
}
\examples{
if (is_kitty()) {
  cat("Yeeh, you are running kitty!")
}

}
