% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitty_colors.R
\name{kitty_colors}
\alias{kitty_colors}
\alias{kitty_background}
\alias{kitty_foreground}
\alias{kitty_palette}
\title{Get the colors used in the kitty terminal}
\usage{
kitty_colors()

kitty_background()

kitty_foreground()

kitty_palette()
}
\value{
\code{kitty_colors} returns a dataframe with the colors from the theme used
by kitty. \code{kitty_background} returns the background color (character
vector with the hex-code). \code{kitty_foreground} returns the foreground
color. \code{kitty_palette} returns a vector with the 9 main accent colors
from the theme.
}
\description{
Get the colors used in the kitty terminal
}
