% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitty_dim.R
\name{kitty_dim}
\alias{kitty_dim}
\title{Get the dimensions of the terminal window in pixels}
\usage{
kitty_dim()
}
\value{
An integer vector with the width and height of the terminal in pixels
(\code{x_pixels} and \code{y_pixels}) and the number of text columns and rows
in the terminal window (\code{columns} and \code{rows}).
}
\description{
Get the dimensions of the terminal window in pixels
}
