% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitty_height.R
\name{kitty_height}
\alias{kitty_height}
\title{Get the height of the terminal window in pixels}
\usage{
kitty_height()
}
\value{
An integer with the number of pixels the terminal is high
}
\description{
Get the height of the terminal window in pixels
}
