% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitty_width.R
\name{kitty_width}
\alias{kitty_width}
\title{Get the with of the terminal window in pixels}
\usage{
kitty_width()
}
\value{
An integer with the number of pixels the terminal is wide.
}
\description{
Get the with of the terminal window in pixels
}
