% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotenv.R
\name{plotenv_begin.kitty}
\alias{plotenv_begin.kitty}
\alias{plotenv_end.kitty}
\title{Functions for use with the plotenv package}
\usage{
plotenv_begin.kitty(name, ...)

plotenv_end.kitty(state)
}
\arguments{
\item{name}{name of the plot. Not used.}

\item{...}{passed on to an internal function. For accepted options see 
\code{\link{kitty_plot}}.}

\item{state}{The state as returned by \code{plotenv_begin.kitty}. This is used
to close the png device and copy the image into the terminal.}
}
\value{
\code{plotenv_begin.kitty} will return some state of the plotting device.
\code{plotenv_end.kitty} does not return anything.
}
\description{
Functions for use with the plotenv package
}
\details{
These functions are not intended to be called directly by the user. When
\code{kitty.r} is loaded and the R-package \code{plotenv} is loaded, these
functions also make a 'kitty' device available for \code{plotenv}. For
example, it becomes possible to use \code{plotenv(device = "kitty", ...)}.
}
