% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png2terminal.R
\name{png2terminal}
\alias{png2terminal}
\title{Dump a PNG image to the terminal using the Terminal Graphics Protocol}
\usage{
png2terminal(filename)
}
\arguments{
\item{filename}{filename of the PNG image}
}
\value{
Called for it's side effect of writing the image to the terminal (standard
out). Return \code{NULL} invisibly.
}
\description{
Dump a PNG image to the terminal using the Terminal Graphics Protocol
}
\details{
The Terminal Graphics Protocol is not supported by many Terminal Emulators.
The most notable terminal emulator supporting the protocol is Kitty.
}
