% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitty_plot.R
\name{kitty_plot}
\alias{kitty_plot}
\title{Create a plot and write the result to the kitty terminal.}
\usage{
kitty_plot(
  expr,
  width = max(480, min(1200, kitty_width(), kitty_height()/0.8)),
  height = 0.8 * width,
  units = "px",
  res = NA,
  ...,
  kitty_col = TRUE,
  kitty_bg = kitty_col,
  kitty_fg = kitty_col
)
}
\arguments{
\item{expr}{Valid R expression creating the plot.}

\item{width}{The width of the image. Passed on to \code{\link{png}}.}

\item{height}{The height of the image. Passed on to \code{\link{png}}.}

\item{units}{The units in which 'height' and 'width' are given. Passed on to
\code{\link{png}}.}

\item{res}{The resulution of the image. Passed on to \code{\link{png}}.}

\item{...}{Passed on to \code{\link{png}}.}

\item{kitty_col}{Logical value indicating that the foreground and background
colors used in the plot should be set to that of the terminal.}

\item{kitty_bg}{Logical value indicating that the background
color used in the plot should be set to that of the terminal.}

\item{kitty_fg}{Logical value indicating that the foreground
color used in the plot should be set to that of the terminal.}
}
\description{
Create a plot and write the result to the kitty terminal.
}
\details{
The \code{kitty_fg} argument sets \code{fg}, \code{col}, \code{col.axis},
\code{col.lab}, \code{col.main} and \code{col.sub} using \code{\link{par}} to
\code{\link{kitty_foreground}}.
}
\examples{
if (is_kitty()) {
  kitty_plot({
    data(iris)
    plot(iris$Sepal.Width, iris$Sepal.Length, 
      col = kitty_palette()[iris$Species], 
      pch = 20)
    grid(lty = 2)
    legend("topright", legend = levels(iris$Species), 
      col = kitty_palette()[1:3], pch = 20)
  })
}

}
